% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{print.fitnmeweibull}
\alias{print.fitnmeweibull}
\title{Print a Fitted NME-Weibull Model}
\usage{
\method{print}{fitnmeweibull}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{"fitnmeweibull"} returned by
\code{\link[=fitnmeweibull]{fitnmeweibull()}}.}

\item{digits}{Number of significant digits to print.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Fitted NME-Weibull Model
}
\examples{
set.seed(1)
print(fitnmeweibull(rnmeweibull(100, 1, 2, 0.5)))
}
