% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmeweibull-package.R
\docType{package}
\name{nmeweibull-package}
\alias{nmeweibull}
\alias{nmeweibull-package}
\title{nmeweibull: New Modified Exponential-Weibull Distribution}
\description{
The nmeweibull package provides functions for the New Modified
Exponential-Weibull distribution. The package includes functions
for density, cumulative distribution, survival, hazard, quantile,
random generation, parameter estimation, and risk measures.
}
\author{
\strong{Maintainer}: Tan Azza Nuro Azzikra \email{222212894@stis.ac.id}

}
\keyword{internal}
